refactor(builder): Move `NodeInputField.tsx` to `node-input.tsx`
