+++ rnd/autogpt_builder/src/components/node-input.tsx
